Changing Age from String to Int in DB
State still treats it like a String, there is still need for treating the string

+++ app/src/main/java/br/com/graest/retinografo/PatientDataViewModel.kt
@@ -60,7 +60,7 @@ class PatientDataViewModel(
                 }
 
                 val patientData = PatientData(
-                    age = age,
+                    age = age.toInt(),
                     name = name
                 )
